requirements.txt
## microservices
fastapi
fastapi[standard]
fastapi-users[beanie]
gradio
uvicorn
pydantic-settings
starlette
redis

# AI and OpenAI integration
openai
azure-identity==1.19.0
azure-mgmt-resource==23.1.1
azure-storage-blob

# HTTPX fix for OpenAI
httpx==0.27.2

# Async and event loop tools
asyncio==3.4.3
tenacity==8.5.0
websockets
pyaudio
python-socks
pipwin
backoff
sounddevice
numpy
dateparser

# python helpers
yfinance
plotly
python-dotenv

# ui
streamlit==1.39.0
streamlit-chat==0.1.1

# ACS
azure-communication-callautomation==1.4.0b1
